requirements.txt
lerobot>=0.4.0
huggingface_hub>=0.20.0
rembg>=2.0.50
opencv-python-headless>=4.8.0
Pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.65.0
numpy>=1.24.0
